tests/wpt/tests/shadow-dom/untriaged/events/test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Shadow DOM Test: A_05_00_01</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#events">
<meta name="assert" content="Events:The mutation event types must never be dispatched in a shadow DOM subtree.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../html/resources/common.js"></script>
</head>
<body>
<div id="log"></div>
<script>
var A_05_00_01_T1 = async_test('A_05_00_01_T1');

A_05_00_01_T1.step(function () {
    var iframe = document.createElement('iframe');
    iframe.src = '../resources/blank.html';
    document.body.appendChild(iframe);

    iframe.onload = A_05_00_01_T1.step_func(function () {

    try {
        var d = iframe.contentDocument;

        var div = d.createElement('div');
        d.body.appendChild(div);

        var s = div.attachShadow({mode: 'open'});

        var div2 = d.createElement('div');
        s.appendChild(div2);

        var inp = d.createElement('input');
        inp.setAttribute('type', 'text');
        inp.setAttribute('id', 'inpid');
        div2.appendChild(inp);

        div2.addEventListener('DOMAttrModified', A_05_00_01_T1.step_func(function (event) {
            assert_true(false, 'The mutation event types must never be dispatched in a shadow DOM subtree');
        }), false);
        /*
        var attr = inp.getAttributeNode ("value");
        var event = d.createEvent('MutationEvent');
        event.initMutationEvent ("DOMAttrModified", true, true, attr, null, 'new value', "value", MutationEvent.MODIFICATION);
        inp.dispatchEvent(event);
            */
        inp.value = 'new value';
        inp.setAttribute ("newAttr" , "firstValue");
        inp.setAttribute ("newAttr" , "secondValue");
        inp.removeAttribute ("newAttr");
    } finally {
        iframe.parentNode.removeChild(iframe);
    }
    A_05_00_01_T1.done();
    });
});
</script>
</body>
</html>
